jobs.runners: move 'SlurmRunner' from 'JobRunner' module.
Relocates and updates the 'SlurmRunner' class from 'JobRunner' into the 'jobs.runners' module, and updates the interface. A version of 'SlurmRunner' which wraps the 'jobs.runners' version is kept in 'JobRunner' for backwards-compatibility.
