+++ bcftbx/jobs/runners.py
@@ -1488,3 +1488,806 @@ class ResourceLock:
             if n == resource_name:
                 return True
         return False
+
+
+class SlurmRunner(JobRunner):
+    """
+    Class implementing job runner for Slurm
+
+    SlurmRunner submits jobs to a Slurm cluster using the ``sbatch``
+    command, determines the status of jobs using ``squeue``, and
+    and terminates them using ``scancel``.
+
+    Additionally the runner can be configured to target a specific
+    partition and number of cores on initialisation.
+
+    Each SlurmRunner instance creates a temporary directory which
+    it uses for internal admin; this will be removed at program
+    exit via 'atexit'.
+
+    Arguments:
+      log_dir (str): path of directory to write log files to (set to 'None'
+        to use cwd)
+      nslots (int): number of threads assigned to the runner instance
+      partition (str): name of Slurm partition to target (set to 'None'
+        to use default queue)
+      join_logs (bool): if True then combine stderr and stdout into a
+        single log file (default is to write stdout and stderr to separate
+        log files)
+      slurm_extra_args (list): arbitrary additional arguments to supply
+        to 'sbatch' (e.g. '["-n", 8]')
+      poll_interval (int): time interval to use (in seconds) when polling
+        Slurm using 'squeue' (seconds) (default: 300.0)
+      timeout (int): maximum length of time to wait (in seconds) before
+        giving up when submitting jobs to Slurm and finalizing jobs
+        (default: 30)
+      missing_job_timeout (int): minimum time (in seconds) that a job
+        needs to be flagged as "missing" before it's removed from the
+        runner (default: 600)
+    """
+
+    def __init__(self, log_dir=None, nslots=None, partition=None,
+                 join_logs=None, slurm_extra_args=None,
+                 poll_interval=300, timeout=30, missing_job_timeout=600):
+        # Internal parameters
+        self._name = "SlurmRunner"
+        self._admin_dir = None
+        self._job_count = 0
+        self._shell = "/bin/bash"
+        # Directory for log files
+        self.set_log_dir(log_dir)
+        # Keep track of data (names, log dirs etc) for each job
+        self._job_number = {}
+        self._names = {}
+        self._log_dirs = {}
+        self._error_state = {}
+        self._exit_status = {}
+        self._finalizing = {}
+        self._queue = {}
+        self._start_time = {}
+        # Job id lock
+        self._job_lock = ResourceLock()
+        # Job grace period lock
+        self._updating_grace_period = ResourceLock()
+        # Job submission lock
+        self._submit_lock = ResourceLock()
+        # Cached job list
+        self._cached_job_list_lifetime = 2.0
+        self._cached_job_list_timestamp = 0.0
+        self._cached_job_list = []
+        self._cached_job_list_force_update = True
+        # Cached qstat output
+        self._cached_squeue_output_lifetime = 0.5
+        self._cached_squeue_output_timestamp = 0.0
+        self._cached_squeue_output = None
+        # Grace period for new jobs
+        self._new_job_grace_period = 2.0
+        # Polling intervals and timeout periods (seconds)
+        self._poll_interval = int(poll_interval)
+        self._timeout = int(timeout)
+        # Handling "missing" jobs (in runner but not in Slurm)
+        self._missing = {}
+        self._missing_job_timeout = int(missing_job_timeout)
+        self._missing_job_last_checked = 0.0
+        # Register clean up function
+        atexit.register(self._clean_up_admin_dir)
+        # Slurm-specific variables
+        self._join_logs = join_logs
+        self._nslots = nslots
+        self._partition = partition
+        if slurm_extra_args is not None:
+            self._slurm_extra_args = [str(x) for x in slurm_extra_args]
+        else:
+            self._slurm_extra_args = None
+        # FIXME clean up extra arguments to remove -n, -p etc?
+        self._check_slurm_extra_args()
+        # Create admin dir
+        self._make_admin_dir()
+
+    def __repr__(self):
+        args = []
+        if self._nslots:
+            args.append(f"nslots={self.nslots}")
+        if self._partition:
+            args.append(f"partition={self._partition}")
+        if self._join_logs:
+            args.append(f"join_logs={self.join_logs}")
+        if self._slurm_extra_args:
+            args.append(f"slurm_args={' '.join(self.slurm_extra_args)}")
+        if args:
+            return self._name + f"({','.join(args)})"
+        else:
+            return self._name
+
+    def name(self, job_id):
+        """
+        Return the name for a specific job
+
+        Arguments:
+          job_id (int): Job ID to get the name for
+        """
+        return self._names[job_id]
+
+    @property
+    def nslots(self):
+        """
+        Return the number of associated slots
+
+        If not set then defaults to 1.
+        """
+        if self._nslots is None:
+            return 1
+        else:
+            return int(self._nslots)
+
+    @property
+    def partition(self):
+        """
+        Return the assigned partition
+
+        If not set then defaults to 1.
+        """
+        return self._partition
+
+    @property
+    def join_logs(self):
+        """
+        Return flag for whether to combine stdout and stderr logs
+
+        If not set then defaults to 'False'.
+        """
+        if self._join_logs is None:
+            return False
+        else:
+            return bool(self._join_logs)
+
+    @property
+    def slurm_extra_args(self):
+        """
+        Return the extra Slurm arguments
+        """
+        if self._slurm_extra_args:
+            return [x for x in self._slurm_extra_args]
+        else:
+            return None
+
+    def run(self, name, working_dir, script, args):
+        """
+        Submit a script or command to the cluster via 'sbatch'
+
+        Arguments:
+          name (str): name to give the job
+          working_dir (str): path to directory to run the job in
+          script (str): path to command or script file to run
+          args (list): list of arguments to supply to the script
+
+        Returns:
+          Job id for submitted job, or 'None' if job failed to
+          start.
+        """
+        logging.debug(f"{self._name:11}: submitting job")
+        logging.debug(f"Name       : {name}")
+        logging.debug(f"Nslots     : {self.nslots}")
+        logging.debug(f"Partition  : {self.partition}")
+        logging.debug(f"Join logs  : {self.join_logs}")
+        logging.debug(f"Extra args : {self.slurm_extra_args}")
+        logging.debug(f"Log dir    : {self.log_dir}")
+        logging.debug(f"Working_dir: {working_dir}")
+        logging.debug(f"Script     : {script}")
+        logging.debug(f"Arguments  : {str(args)}")
+        # Wait for lock on job submission
+        start_time = time.time()
+        submit_lock = None
+        while submit_lock is None:
+            submit_lock = self._submit_lock.acquire("job_submission",
+                                                    timeout=self._timeout)
+        # Get internal job number
+        self._job_count += 1
+        job_number = self._job_count
+        logging.debug("Internal job count: %s" % job_number)
+        # Release the lock
+        self._submit_lock.release(submit_lock)
+        # Build script to run the command to be submitted
+        job_dir = os.path.join(self._admin_dir, str(job_number))
+        logging.debug("Job admin dir     : %s" % job_dir)
+        os.mkdir(job_dir)
+        cmd_args = [script]
+        for arg in args:
+            # Quote arguments containing whitespace
+            if arg.count(' ') or arg.count('\t'):
+                arg = "\"%s\"" % arg
+            cmd_args.append(arg)
+        cmd = ' '.join(cmd_args)
+        job_script = os.path.join(job_dir, "job_script.sh")
+        with open(job_script, "wt") as fp:
+            fp.write(u"""#!{shell}
+export BCFTBX_RUNNER_NSLOTS=$SLURM_NTASKS
+echo "$BCFTBX_RUNNER_NSLOTS" > {job_dir}/__jobrunner_nslots
+{cmd}
+exit_code=$?
+echo "$exit_code" > {job_dir}/__exit_code.tmp
+mv {job_dir}/__exit_code.tmp {job_dir}/__exit_code
+exit $exit_code
+""".format(shell=self._shell, job_dir=job_dir, cmd=cmd))
+        os.chmod(job_script,0o755)
+        job_name = self._sanitize_job_name(name)
+        logging.debug("Slurm job name: %s" % job_name)
+        # Log file (replicate Grid Engine)
+        stdout = "%x.o%j"
+        if self.log_dir:
+            stdout = os.path.join(self.log_dir, stdout)
+        if self.join_logs:
+            stderr = None
+        else:
+            stderr = "%x.e%j"
+            if self.log_dir:
+                stderr = os.path.join(self.log_dir, stderr)
+        # Build sbatch command to submit script
+        sbatch = ["sbatch",
+                  "--export=ALL",
+                  "-J", job_name,
+                  "-n", str(self.nslots)]
+        if self._partition:
+            sbatch.extend(["-p", self._partition])
+        sbatch.extend(["-o", stdout])
+        if stderr:
+            sbatch.extend(["-e", stderr])
+        if working_dir:
+            sbatch.extend(('--chdir',working_dir))
+        if self.slurm_extra_args:
+            sbatch.extend(self.slurm_extra_args)
+        sbatch.append(job_script)
+        logging.debug("SlurmRunner: sbatch command: %s" % sbatch)
+        # Run the sbatch job in the current directory
+        cwd = os.getcwd()
+        if not os.path.exists(cwd):
+            logging.error("SlurmRunner: cwd doesn't exist!")
+            return None
+        logging.debug("SlurmRunner: executing in %s" % cwd)
+        p = subprocess.Popen(sbatch, cwd=cwd,
+                             stdout=subprocess.PIPE,
+                             stderr=subprocess.PIPE,
+                             universal_newlines=True)
+        stdoutdata, stderrdata = p.communicate()
+        logging.debug(f"SlurmRunner: sbatch output: {stdoutdata}")
+        logging.debug(f"SlurmRunner: sbatch error: {stderrdata}")
+        # Check stderr to try and detect error with submission
+        error = stderrdata.strip()
+        if error:
+            # Just echo error message as a warning
+            logging.warning("SlurmRunner: '%s'" % error)
+        # Capture the job id from the output
+        job_id = None
+        for line in stdoutdata.split('\n'):
+            if line.startswith("Submitted batch job"):
+                job_id = line.split()[-1]
+        if job_id is None:
+            logging.error("SlurmRunner: failed to get job ID from "
+                          "sbatch output: %r" % stdoutdata)
+        logging.debug(f"SlurmRunner: done - job id = {job_id}")
+        # Store internal number, name and log dir against job id
+        if job_id is not None:
+            self._job_number[job_id] = job_number
+            self._names[job_id] = name
+            if self.log_dir is None:
+                self._log_dirs[job_id] = working_dir
+            else:
+                self._log_dirs[job_id] = self.log_dir
+            self._start_time[job_id] = time.time()
+        # Force refresh of job list
+        self._cached_job_list_force_update = True
+        # Return the job id
+        return job_id
+
+    def terminate(self, job_id, exit_code=-1):
+        """
+        Remove a job from the Slurm queue using 'scancel'
+        """
+        logging.debug("SlurmRunner: deleting job")
+        scancel=("scancel", job_id)
+        p = subprocess.Popen(scancel, stdout=subprocess.PIPE)
+        stdoutdata, stderrdata = p.communicate()
+        message = stdoutdata.strip()
+        logging.debug("Slurmrunner: scancel: %s" % message)
+        if job_id in self._start_time:
+            del(self._start_time[job_id])
+        # Write an exit code file for the job
+        exit_code_file = os.path.join(self._admin_dir,
+                                      str(self._job_number[job_id]),
+                                      "__exit_code")
+        with open("%s.tmp" % exit_code_file, "wt") as fp:
+            fp.write(f"{exit_code}\n")
+        os.rename("%s.tmp" % exit_code_file, exit_code_file)
+        # Force update of cached job list
+        self._cached_job_list_force_update = True
+        return True
+
+    def log_file(self, job_id):
+        """
+        Return the log file name for a job
+
+        The name should be '<name>.o<job_id>'
+        """
+        name = self._names[job_id]
+        log_file = "%s.o%s" % (name, job_id)
+        if self._log_dirs[job_id] is not None:
+            log_file = os.path.join(self._log_dirs[job_id], log_file)
+        return log_file
+
+    def err_file(self, job_id):
+        """
+        Return the error file name for a job
+
+        The name should be '<name>.e<job_id>'
+        """
+        if self.join_logs:
+            return self.log_file(job_id)
+        name = self._names[job_id]
+        err_file = "%s.e%s" % (name,job_id)
+        if self._log_dirs[job_id] is not None:
+            err_file = os.path.join(self._log_dirs[job_id], err_file)
+        return err_file
+
+    def error_state(self,job_id):
+        """
+        Check if the job is in an error state
+
+        Return True if the job is deemed to be in an 'error
+        state', False otherwise.
+        """
+        logging.debug("SlurmRunner: 'error_state' method not implemented")
+        return False
+
+    def list(self):
+        """
+        Get list of job ids which are queued or running
+        """
+        # Check cached job list
+        use_cache = (not self._cached_job_list_force_update and
+                     (time.time() - self._cached_job_list_timestamp) <
+                     self._cached_job_list_lifetime)
+        if use_cache:
+            logging.debug("SlurmRunner: using cached job list")
+            job_ids = self._cached_job_list
+            # Add the jobs in grace period
+            for job_id in self._grace_period_jobs():
+                if job_id not in job_ids:
+                    job_ids.append(job_id)
+            return job_ids
+        else:
+            logging.debug("SlurmRunner: building job list")
+        # Update jobs in grace period
+        for job_id in self._grace_period_jobs():
+            self._update_job_in_grace_period(job_id)
+        # Build initial list from directory contents
+        job_ids = []
+        for job_id in list(self._job_number.keys()):
+            logging.debug(f"SlurmRunner: -- checking job {job_id}")
+            try:
+                job_number = self._job_number[job_id]
+            except KeyError:
+                # Job has been removed since the list was
+                # fetched? Ignore
+                continue
+            job_dir = os.path.join(self._admin_dir, str(job_number))
+            if os.path.exists(job_dir):
+                # Job dir exists
+                logging.debug("SlurmRunner: -- found %s" % job_dir)
+                exit_code_file = os.path.join(job_dir, "__exit_code")
+                if os.path.exists(exit_code_file):
+                    # Job has finished, handle completion
+                    logging.debug("SlurmRunner: -- exit code file exists, "
+                                  f"completing job {job_id}")
+                    self._handle_job_completion(job_id)
+                else:
+                    # Job still running
+                    logging.debug("SlurmRunner: -- no exit code file, "
+                                  f"{job_id} still running")
+                    job_ids.append(job_id)
+        # Check for "missing" jobs that are in the runner but no
+        # longer in the Slurm system
+        jobs_still_in_grace_period = bool(self._grace_period_jobs())
+        if job_ids and not jobs_still_in_grace_period:
+            check_missing_jobs = ((time.time() -
+                                   self._missing_job_last_checked) >
+                                  self._poll_interval)
+            if check_missing_jobs:
+                logging.debug(f"SlurmRunner: checking for missing jobs")
+                job_ids = self._handle_missing_jobs(job_ids)
+                self._missing_job_last_checked = time.time()
+        # Update cache
+        logging.debug("SlurmRunner: updating the cache")
+        self._cached_job_list_timestamp = time.time()
+        self._cached_job_list = [j for j in job_ids]
+        self._cached_job_list_force_update = False
+        # Add the jobs in the grace period
+        for job_id in self._grace_period_jobs():
+            if job_id not in job_ids:
+                job_ids.append(job_id)
+            else:
+                # Job now visible so no longer in grace period
+                self._update_job_in_grace_period(job_id)
+        logging.debug("SlurmRunner: 'list' returning %s" % job_ids)
+        return job_ids
+
+    def exit_status(self,job_id):
+        """
+        Return exit status from command run by a job
+
+        If the job is still running then returns 'None'.
+        """
+        if self.is_running(job_id):
+            # Return None if job is still running
+            return None
+        # Check if job is being finalized
+        start_time = time.time()
+        while job_id in self._finalizing:
+            # Wait until exit_status is ready
+            time.sleep(1.0)
+            if (time.time() - start_time) > self._timeout:
+                logging.warning("SlurmRunner: timed out waiting "
+                                "for job %s to finalize" % job_id)
+                return None
+        # Return cached exit status
+        return self._exit_status[job_id]
+
+    def _make_admin_dir(self):
+        """
+        Internal: create temporary directory for admin
+
+        The directory will be created in a subdirectory of
+        the current working directory.
+        """
+        if self._admin_dir:
+            # Return current value, if set
+            return self._admin_dir
+        # Create a new temporary directory
+        self._admin_dir = tempfile.mkdtemp(prefix=".slurmrunner.",
+                                           dir=os.getcwd())
+        return self._admin_dir
+
+    def _clean_up_admin_dir(self):
+        """
+        Internal: remove the admin dir
+
+        Shouldn't be called directly; instead register with
+        'atexit' to force clean up on program exit
+        """
+        logging.debug("SlurmRunner: removing admin dir '%s'" %
+                      self._admin_dir)
+        # Check if jobs are still being finalized
+        start_time = time.time()
+        while self._finalizing:
+            # Wait until everything has finalized
+            time.sleep(1.0)
+            if (time.time() - start_time) > self._timeout:
+                logging.warning("SlurmRunner: timed out waiting "
+                                "for jobs to finalize")
+                break
+        # Try to remove the admin dir and contents
+        try:
+            shutil.rmtree(self._admin_dir)
+        except Exception as ex:
+            logging.warning("SlurmRunner: exception removing "
+                            "admin dir '%s': %s" %
+                            (self._admin_dir, ex))
+
+    def _grace_period_jobs(self):
+        """
+        Internal: return list of jobs in the grace period
+
+        Returns list of job IDs where each job has an entry
+        in the `_start_time` dictionary which is newer than
+        the grace period timeout.
+        """
+        return list(self._start_time.keys())
+
+    def _update_job_in_grace_period(self,job_id):
+        """
+        Internal: handling update of job in grace period
+
+        Checks if a job is still within the grace period
+        (i.e. has an entry in the `_start_time`
+        dictionary which is newer than the grace period).
+
+        If the job is no longer in the grace period then
+        removes its entry in the `_start_time`
+        dictionary.
+        """
+        logging.debug("SlurmRunner: checking if job %s is still in "
+                      "grace period" % job_id)
+        lock = None
+        while lock is None:
+            lock = self._updating_grace_period.acquire(job_id)
+        logging.debug("SlurmRunner: acquired lock for grace period "
+                      "update: %s" % lock)
+        try:
+            start_time = self._start_time[job_id]
+        except KeyError:
+            logging.debug("SlurmRunner: update grace period: job %s "
+                          "has gone away (ignored)" % job_id)
+            self._updating_grace_period.release(lock)
+            return
+        if ((time.time() - start_time) > self._new_job_grace_period):
+            # Job no longer in grace period
+            logging.debug("SlurmRunner: job %s no longer in grace "
+                          "period" % job_id)
+            try:
+                del(self._start_time[job_id])
+            except KeyError:
+                logging.debug("SlurmRunner: update grace period: "
+                              "job %s has gone away (ignored)" %
+                              job_id)
+        # Release update lock
+        self._updating_grace_period.release(lock)
+
+    def _handle_missing_jobs(self, job_list):
+        """
+        Internal: handle any jobs in runner missing from Slurm
+
+        Jobs present in the runner are checked to see if they
+        are also present in Slurm, with jobs that are missing
+        being flagged initially and then removed from the runner
+        if still missing after a "timeout" period.
+
+        Arguments:
+          job_list (list): list of job IDs to check
+
+        Returns:
+          List: updated list of job IDs with missing jobs
+            removed.
+        """
+        updated_job_list = []
+        squeue_jobs = [j[0] for j in self._run_squeue()]
+        for job_id in job_list:
+            if job_id not in squeue_jobs:
+                logging.debug(f"SlurmRunner: job {job_id} has gone away?")
+                if job_id not in self._missing:
+                    # Set time when job went missing
+                    self._missing[job_id] = time.time()
+                    updated_job_list.append(job_id)
+                elif (time.time() - self._missing[job_id]) > \
+                     self._missing_job_timeout:
+                    # Job is still missing after interval, terminate it
+                    logging.debug(f"SlurmRunner: forcing job completion "
+                                  f"for missing job {job_id} from runner "
+                                  f"(timeout was exceeded)")
+                    self.terminate(job_id, exit_code=127)
+                    # Perform immediate job completion
+                    self._handle_job_completion(job_id)
+                    # Remove the "missing" flag
+                    del(self._missing[job_id])
+            else:
+                # Job is no longer missing?
+                logging.debug(f"SlurmRunner: previously missing job {job_id} "
+                              f"has come back?")
+                updated_job_list.append(job_id)
+                if job_id in self._missing:
+                    del(self._missing[job_id])
+        return updated_job_list
+
+    def _handle_job_completion(self,job_id):
+        """
+        Internal: deal with completion of job
+
+        Peforms the following operations:
+
+        - checks that an '__exit_code' file exists for
+          the job
+        - read and store the exit status/return code from
+          this file
+        - ensure that the queue is set for the job
+        - call the clean up function to remove all the
+          associated files
+        - remove the job from the internal job count
+
+        If the '__exit_code' file associated with the job
+        can't be found after a number of attempts to locate
+        it, or if the exit status cannot be read from the
+        file, then the exit status for the job will be
+        set to '127'.
+        """
+        logging.debug("SlurmRunner: handle job completion for %s"
+                      % job_id)
+        lock = None
+        while lock is None:
+            lock = self._job_lock.acquire(job_id)
+        logging.debug("SlurmRunner: acquired lock: %s" % lock)
+        if job_id not in self._job_number:
+            # Job has gone away
+            logging.debug("SlurmRunner: job %s has gone away" %
+                          job_id)
+            self._job_lock.release(lock)
+            return
+        self._finalizing[job_id] = True
+        # Check there is an exit code file
+        exit_code_file = os.path.join(self._admin_dir,
+                                      str(self._job_number[job_id]),
+                                      "__exit_code")
+        assert(os.path.exists(exit_code_file))
+        try:
+            with open(exit_code_file,'rt') as fp:
+                exit_status = int(fp.read())
+        except Exception as ex:
+            # Set exit status to 127
+            logging.error("SlurmRunner: exception when "
+                          "reading exit_status for job "
+                          "%s: %s" % (job_id, ex))
+            exit_status = 127
+        # Store exit status and clean up
+        self._exit_status[job_id] = exit_status
+        self._clean_up_job(job_id)
+        # Release finalization lock
+        del(self._finalizing[job_id])
+        # Release job lock
+        self._job_lock.release(lock)
+
+    def _clean_up_job(self,job_id):
+        """
+        Internal: clean up internal job files
+
+        Removes the internal directory associated with a job, along
+        with any files it contains (e.g. job script, exit code etc).
+
+        This method should only be invoked for jobs that have
+        finished running. If the job is still running then returns
+        with no action.
+        """
+        # Do clean up
+        logging.debug("SlurmRunner: cleaning up after job %s" % job_id)
+        try:
+            job_number = self._job_number[job_id]
+        except KeyError:
+            logging.error("SlurmRunner: job %d not found, can't do "
+                          "clean up" % job_id)
+            return
+        job_dir = os.path.join(self._admin_dir, str(job_number))
+        try:
+            # Remove the directory and contents
+            shutil.rmtree(job_dir)
+        except Exception as ex:
+            logging.warning("SlurmRunner: exception cleaning up for "
+                            "job %s (ignored): %s" % (job_id, ex))
+        # Clear stored error state
+        try:
+            del(self._error_state[job_id])
+        except KeyError:
+            pass
+        # Remove the internally stored job number
+        del(self._job_number[job_id])
+
+    def _run_squeue(self):
+        """
+        Internal: run squeue and return data as a list of lists
+
+        Runs 'squeue' command, processes the output and returns a
+        list where each item is the data for a job in the form of
+        another list, with the items in this list being the data
+        returned by squeue.
+
+        NB as 'squeue' calls can be relatively expensive to make,
+        the output from 'squeue' is cached for a specified period
+        before being refreshed.
+        """
+        # Should we return the cached data?
+        if (time.time() - self._cached_squeue_output_timestamp) < \
+           self._cached_squeue_output_lifetime:
+            logging.debug("SlurmRunner: returning cached squeue output")
+            return self._cached_squeue_output
+        # Run squeue and collect the output
+        try:
+            cmd = ["squeue", "--user", os.getlogin()]
+        except OSError:
+            # os.getlogin() not guaranteed to work in all environments?
+            cmd = ["squeue", "--me"]
+        # Run squeue command
+        p = subprocess.Popen(cmd,
+                             stdout=subprocess.PIPE,
+                             universal_newlines=True)
+        stdoutdata = p.communicate()[0]
+        logging.debug(f"SlurmRunner: output from 'squeue': {stdoutdata}")
+        # Process the output
+        squeue_output = []
+        # Output has a header line with field names then one
+        # line per job
+        lines = stdoutdata.rstrip("\n").split("\n")
+        fields = lines[0].split()
+        idx_jobid = fields.index("JOBID")
+        idx_state = fields.index("ST")
+        for line in lines[1:]:
+            # Store tuples of jobid, state
+            data = line.split()
+            try:
+                squeue_output.append((data[idx_jobid], data[idx_state]))
+            except IndexError:
+                logging.debug(f"SlurmRunner: failed to parse 'squeue' "
+                              f"output: '{line}' (ignored)")
+        # Update the cache
+        self._cached_squeue_output_timestamp = time.time()
+        self._cached_squeue_output = squeue_output
+        return squeue_output
+
+    def _job_state_code(self, job_id):
+        """
+        Internal: get the state code for a job id
+
+        Will be one of the Slurm job state codes, or an empty
+        string if the job id isn't found.
+        """
+        # Run squeue and process output to get job states
+        logging.debug("SlurmRunner: acquiring state for job %s"
+                      % job_id)
+        squeue = self._run_squeue()
+        job_ids = []
+        job_states = {}
+        for job_data in squeue:
+            id_ = job_data[0]
+            state = job_data[1]
+            logging.debug(f"SlurmRunner: found job {id_} (state '{state}')")
+            if id_ == job_id:
+                return state
+        # Job not found
+        return ""
+
+    def _sanitize_job_name(self, name):
+        """
+        Internal: sanitize a job name for use with Slurm
+
+        This is copied from the GE runner
+        """
+        sanitized_name = str(name)
+        for c in ":*@\\?":
+            sanitized_name = sanitized_name.replace(c,'_')
+        if sanitized_name[0].isdigit():
+            # If name starts with a digit then prepend
+            # an underscore
+            sanitized_name = f"_{sanitized_name}"
+        return sanitized_name
+
+    def _check_slurm_extra_args(self):
+        """
+        Internal: check extra arguments provided for Slurm
+
+        The -n and -p options implicitly set the number of
+        slots and the target partition, but only if not
+        explicitly set when the runner was instantiated
+        (otherwise an exception is raised).
+
+        The extra arguments list is updated to remove the
+        explicit -n and -p options.
+
+        The -J, -o, -e and --export options are reserved for
+        exclusive use by the runner so if these are specified
+        then an exception is raised.
+        """
+        if self._slurm_extra_args is None:
+            return
+        args = [x for x in self._slurm_extra_args]
+        new_args = []
+        while args:
+            # Grab first argument and update the list
+            arg = args[0]
+            args = args[1:]
+            # Check for specific arguments
+            if arg == "-n":
+                if self._nslots is None:
+                    self._nslots = int(args[0])
+                    args = args[1:]
+                else:
+                    raise Exception(f"SlurmRunner: '-n' not permitted "
+                                    f"in extra arguments if nslots is set")
+            elif arg == "-p":
+                if self._partition is None:
+                    self._partition = str(args[0])
+                    args = args[1:]
+                else:
+                    raise Exception(f"SlurmRunner: '-p' not permitted "
+                                    f"in extra arguments if partition is set")
+            elif arg in ("-J", "-o", "-e") or \
+                 arg == "--export" or arg.startswith("--export="):
+                raise Exception(f"SlurmRunner: '{arg}' not permitted "
+                                f"in extra arguments")
+            else:
+                new_args.append(arg)
+        # Update the extra arguments
+        self._slurm_extra_args = new_args
